[utils/viewer] move viewer code out of esp/gfx library
