--- src/esp/gfx/Viewer.cpp
@@ -1,428 +1,0 @@
-// Copyright (c) Facebook, Inc. and its affiliates.
-// This source code is licensed under the MIT license found in the
-// LICENSE file in the root directory of this source tree.
-
-#include <stdlib.h>
-
-#include "Viewer.h"
-
-#include <Corrade/Utility/Arguments.h>
-#include <Corrade/Utility/Directory.h>
-#include <Magnum/DebugTools/Screenshot.h>
-#include <Magnum/EigenIntegration/GeometryIntegration.h>
-#include <Magnum/GL/DefaultFramebuffer.h>
-#include <Magnum/GL/Renderer.h>
-#include <sophus/so3.hpp>
-#include "Drawable.h"
-#include "esp/core/esp.h"
-#include "esp/io/io.h"
-
-#include "esp/gfx/Simulator.h"
-#include "esp/scene/SceneConfiguration.h"
-
-#include "esp/gfx/configure.h"
-
-using namespace Magnum;
-using namespace Math::Literals;
-using namespace Corrade;
-
-constexpr float moveSensitivity = 0.1f;
-constexpr float lookSensitivity = 11.25f;
-constexpr float rgbSensorHeight = 1.5f;
-
-namespace esp {
-namespace gfx {
-
-Viewer::Viewer(const Arguments& arguments)
-    : Platform::Application{arguments,
-                            Configuration{}.setTitle("Viewer").setWindowFlags(
-                                Configuration::WindowFlag::Resizable),
-                            GLConfiguration{}
-                                .setColorBufferSize(Vector4i(8, 8, 8, 8))
-                                .setSampleCount(4)},
-      pathfinder_(nav::PathFinder::create()),
-      controls_(),
-      previousPosition_() {
-  Utility::Arguments args;
-#ifdef CORRADE_TARGET_EMSCRIPTEN
-  args.addNamedArgument("scene")
-#else
-  args.addArgument("scene")
-#endif
-      .setHelp("scene", "scene file to load")
-      .addSkippedPrefix("magnum", "engine-specific options")
-      .setGlobalHelp("Displays a 3D scene file provided on command line")
-      .addBooleanOption("enable-physics")
-      .addOption("physics-config", ESP_DEFAULT_PHYS_SCENE_CONFIG)
-      .setHelp("physics-config", "physics scene config file")
-      .parse(arguments.argc, arguments.argv);
-
-  const auto viewportSize = GL::defaultFramebuffer.viewport().size();
-  enablePhysics_ = args.isSet("enable-physics");
-  std::string physicsConfigFilename = args.value("physics-config");
-  if (!Utility::Directory::exists(physicsConfigFilename)) {
-    LOG(ERROR)
-        << physicsConfigFilename
-        << " was not found, specify an existing file in --physics-config";
-    std::exit(1);
-  }
-
-  // Setup renderer and shader defaults
-  GL::Renderer::enable(GL::Renderer::Feature::DepthTest);
-  GL::Renderer::enable(GL::Renderer::Feature::FaceCulling);
-
-  int sceneID = sceneManager_.initSceneGraph();
-  sceneID_.push_back(sceneID);
-  sceneGraph_ = &sceneManager_.getSceneGraph(sceneID);
-  rootNode_ = &sceneGraph_->getRootNode();
-  navSceneNode_ = &rootNode_->createChild();
-
-  auto& drawables = sceneGraph_->getDrawables();
-  const std::string& file = args.value("scene");
-  const assets::AssetInfo info = assets::AssetInfo::fromPath(file);
-
-  if (enablePhysics_) {
-    if (!resourceManager_.loadScene(info, physicsManager_, navSceneNode_,
-                                    &drawables, physicsConfigFilename)) {
-      LOG(ERROR) << "cannot load " << file;
-      std::exit(1);
-    }
-  } else {
-    if (!resourceManager_.loadScene(info, navSceneNode_, &drawables)) {
-      LOG(ERROR) << "cannot load " << file;
-      std::exit(1);
-    }
-  }
-
-  // Set up camera
-  renderCamera_ = &sceneGraph_->getDefaultRenderCamera();
-  agentBodyNode_ = &rootNode_->createChild();
-  rgbSensorNode_ = &agentBodyNode_->createChild();
-
-  rgbSensorNode_->translate({0.0f, rgbSensorHeight, 0.0f});
-  agentBodyNode_->translate({0.0f, 0.0f, 5.0f});
-
-  float hfov = 90.0f;
-  int width = viewportSize[0];
-  int height = viewportSize[1];
-  float znear = 0.01f;
-  float zfar = 1000.0f;
-  renderCamera_->setProjectionMatrix(width, height, znear, zfar, hfov);
-
-  // Load navmesh if available
-  const std::string navmeshFilename = io::changeExtension(file, ".navmesh");
-  if (io::exists(navmeshFilename)) {
-    LOG(INFO) << "Loading navmesh from " << navmeshFilename;
-    pathfinder_->loadNavMesh(navmeshFilename);
-
-    const vec3f position = pathfinder_->getRandomNavigablePoint();
-    agentBodyNode_->setTranslation(Vector3(position));
-  }
-
-  // connect controls to navmesh if loaded
-  if (pathfinder_->isLoaded()) {
-    controls_.setMoveFilterFunction([&](const vec3f& start, const vec3f& end) {
-      vec3f currentPosition = pathfinder_->tryStep(start, end);
-      LOG(INFO) << "position=" << currentPosition.transpose() << " rotation="
-                << quatf(agentBodyNode_->rotation()).coeffs().transpose();
-      LOG(INFO) << "Distance to closest obstacle: "
-                << pathfinder_->distanceToClosestObstacle(currentPosition);
-
-      return currentPosition;
-    });
-  }
-
-  renderCamera_->node().setTransformation(
-      rgbSensorNode_->absoluteTransformation());
-
-  timeline_.start();
-
-}  // end Viewer::Viewer
-
-void Viewer::addObject(std::string configFile) {
-  if (physicsManager_ == nullptr)
-    return;
-
-  Magnum::Matrix4 T =
-      agentBodyNode_
-          ->MagnumObject::transformationMatrix();  // Relative to agent bodynode
-  Vector3 new_pos = T.transformPoint({0.1f, 2.5f, -2.0f});
-
-  auto& drawables = sceneGraph_->getDrawables();
-  assets::PhysicsObjectAttributes poa =
-      resourceManager_.getPhysicsObjectAttributes(configFile);
-  int physObjectID = physicsManager_->addObject(configFile, &drawables);
-  physicsManager_->setTranslation(physObjectID, new_pos);
-
-  // draw random quaternion via the method:
-  // http://planning.cs.uiuc.edu/node198.html
-  double u1 = (rand() % 1000) / 1000.0;
-  double u2 = (rand() % 1000) / 1000.0;
-  double u3 = (rand() % 1000) / 1000.0;
-
-  Magnum::Vector3 qAxis(sqrt(1 - u1) * cos(2 * M_PI * u2),
-                        sqrt(u1) * sin(2 * M_PI * u3),
-                        sqrt(u1) * cos(2 * M_PI * u3));
-  physicsManager_->setRotation(
-      physObjectID,
-      Magnum::Quaternion(qAxis, sqrt(1 - u1) * sin(2 * M_PI * u2)));
-  objectIDs_.push_back(physObjectID);
-
-  // const Magnum::Vector3& trans =
-  // physicsManager_->getTranslation(physObjectID); LOG(INFO) << "translation: "
-  // << trans[0] << ", " << trans[1] << ", " << trans[2];
-}
-
-void Viewer::removeLastObject() {
-  if (physicsManager_ == nullptr || objectIDs_.size() == 0)
-    return;
-  physicsManager_->removeObject(objectIDs_.back());
-  objectIDs_.pop_back();
-}
-
-void Viewer::invertGravity() {
-  if (physicsManager_ == nullptr)
-    return;
-  const Magnum::Vector3& gravity = physicsManager_->getGravity();
-  const Magnum::Vector3 invGravity = -1 * gravity;
-  physicsManager_->setGravity(invGravity);
-}
-
-void Viewer::pokeLastObject() {
-  if (physicsManager_ == nullptr || objectIDs_.size() == 0)
-    return;
-  Magnum::Matrix4 T =
-      agentBodyNode_
-          ->MagnumObject::transformationMatrix();  // Relative to agent bodynode
-  Vector3 impulse = T.transformPoint({0.0f, 0.0f, -3.0f});
-  Vector3 rel_pos = Vector3(0.0f, 0.0f, 0.0f);
-  physicsManager_->applyImpulse(objectIDs_.back(), impulse, rel_pos);
-}
-
-void Viewer::pushLastObject() {
-  if (physicsManager_ == nullptr || objectIDs_.size() == 0)
-    return;
-  Magnum::Matrix4 T =
-      agentBodyNode_
-          ->MagnumObject::transformationMatrix();  // Relative to agent bodynode
-  Vector3 force = T.transformPoint({0.0f, 0.0f, -40.0f});
-  Vector3 rel_pos = Vector3(0.0f, 0.0f, 0.0f);
-  physicsManager_->applyForce(objectIDs_.back(), force, rel_pos);
-}
-
-void Viewer::torqueLastObject() {
-  if (physicsManager_ == nullptr || objectIDs_.size() == 0)
-    return;
-  Vector3 torque = randomDirection() * 30;
-  physicsManager_->applyTorque(objectIDs_.back(), torque);
-}
-
-// generate random direction vectors:
-Magnum::Vector3 Viewer::randomDirection() {
-  Magnum::Vector3 dir(1.0f, 1.0f, 1.0f);
-  while (sqrt(dir.dot()) > 1.0) {
-    dir = Magnum::Vector3((float)((rand() % 2000 - 1000) / 1000.0),
-                          (float)((rand() % 2000 - 1000) / 1000.0),
-                          (float)((rand() % 2000 - 1000) / 1000.0));
-  }
-  dir = dir / sqrt(dir.dot());
-  return dir;
-}
-
-void Viewer::wiggleLastObject() {
-  // demo of kinematic motion capability
-  // randomly translate last added object
-  if (physicsManager_ == nullptr || objectIDs_.size() == 0)
-    return;
-
-  physicsManager_->translate(objectIDs_.back(), randomDirection() * 0.1);
-}
-
-Vector3 Viewer::positionOnSphere(Magnum::SceneGraph::Camera3D& camera,
-                                 const Vector2i& position) {
-  // Convert from window to frame coordinates.
-  Vector2 framePosition =
-      (Vector2{position} * Vector2{framebufferSize()}) / Vector2{windowSize()};
-  const Vector2 positionNormalized =
-      framePosition / Vector2{camera.viewport()} - Vector2{0.5f};
-  const Float length = positionNormalized.length();
-  const Vector3 result(length > 1.0f
-                           ? Vector3(positionNormalized, 0.0f)
-                           : Vector3(positionNormalized, 1.0f - length));
-  return (result * Vector3::yScale(-1.0f)).normalized();
-}
-
-void Viewer::drawEvent() {
-  GL::defaultFramebuffer.clear(GL::FramebufferClear::Color |
-                               GL::FramebufferClear::Depth);
-  if (sceneID_.size() <= 0)
-    return;
-
-  if (physicsManager_ != nullptr)
-    physicsManager_->stepPhysics(timeline_.previousFrameDuration());
-
-  int DEFAULT_SCENE = 0;
-  int sceneID = sceneID_[DEFAULT_SCENE];
-  auto& sceneGraph = sceneManager_.getSceneGraph(sceneID);
-  renderCamera_->getMagnumCamera().draw(sceneGraph.getDrawables());
-
-  swapBuffers();
-  timeline_.nextFrame();
-  redraw();
-}
-
-void Viewer::viewportEvent(ViewportEvent& event) {
-  GL::defaultFramebuffer.setViewport({{}, framebufferSize()});
-  renderCamera_->getMagnumCamera().setViewport(event.windowSize());
-}
-
-void Viewer::mousePressEvent(MouseEvent& event) {
-  if (event.button() == MouseEvent::Button::Left)
-    previousPosition_ =
-        positionOnSphere(renderCamera_->getMagnumCamera(), event.position());
-
-  event.setAccepted();
-}
-
-void Viewer::mouseReleaseEvent(MouseEvent& event) {
-  if (event.button() == MouseEvent::Button::Left)
-    previousPosition_ = Vector3();
-
-  event.setAccepted();
-}
-
-void Viewer::mouseScrollEvent(MouseScrollEvent& event) {
-  if (!event.offset().y()) {
-    return;
-  }
-
-  /* Distance to origin */
-  const float distance =
-      renderCamera_->node().transformation().translation().z();
-
-  /* Move 15% of the distance back or forward */
-  renderCamera_->node().translateLocal(
-      {0.0f, 0.0f,
-       distance * (1.0f - (event.offset().y() > 0 ? 1 / 0.85f : 0.85f))});
-
-  event.setAccepted();
-}
-
-void Viewer::mouseMoveEvent(MouseMoveEvent& event) {
-  if (!(event.buttons() & MouseMoveEvent::Button::Left)) {
-    return;
-  }
-
-  const Vector3 currentPosition =
-      positionOnSphere(renderCamera_->getMagnumCamera(), event.position());
-  const Vector3 axis = Math::cross(previousPosition_, currentPosition);
-
-  if (previousPosition_.length() < 0.001f || axis.length() < 0.001f) {
-    return;
-  }
-  const auto angle = Math::angle(previousPosition_, currentPosition);
-  renderCamera_->node().rotate(-angle, axis.normalized());
-  previousPosition_ = currentPosition;
-
-  event.setAccepted();
-}
-
-void Viewer::keyPressEvent(KeyEvent& event) {
-  const auto key = event.key();
-  switch (key) {
-    case KeyEvent::Key::Esc:
-      std::exit(0);
-      break;
-    case KeyEvent::Key::Left:
-      controls_(*agentBodyNode_, "lookLeft", lookSensitivity);
-      break;
-    case KeyEvent::Key::Right:
-      controls_(*agentBodyNode_, "lookRight", lookSensitivity);
-      break;
-    case KeyEvent::Key::Up:
-      controls_(*rgbSensorNode_, "lookUp", lookSensitivity, false);
-      break;
-    case KeyEvent::Key::Down:
-      controls_(*rgbSensorNode_, "lookDown", lookSensitivity, false);
-      break;
-    case KeyEvent::Key::Nine:
-      if (pathfinder_->isLoaded()) {
-        const vec3f position = pathfinder_->getRandomNavigablePoint();
-        agentBodyNode_->setTranslation(Vector3(position));
-      }
-      break;
-    case KeyEvent::Key::A:
-      controls_(*agentBodyNode_, "moveLeft", moveSensitivity);
-      LOG(INFO) << "Agent position "
-                << Eigen::Map<vec3f>(agentBodyNode_->translation().data());
-      break;
-    case KeyEvent::Key::D:
-      controls_(*agentBodyNode_, "moveRight", moveSensitivity);
-      LOG(INFO) << "Agent position "
-                << Eigen::Map<vec3f>(agentBodyNode_->translation().data());
-      break;
-    case KeyEvent::Key::S:
-      controls_(*agentBodyNode_, "moveBackward", moveSensitivity);
-      LOG(INFO) << "Agent position "
-                << Eigen::Map<vec3f>(agentBodyNode_->translation().data());
-      break;
-    case KeyEvent::Key::W:
-      controls_(*agentBodyNode_, "moveForward", moveSensitivity);
-      LOG(INFO) << "Agent position "
-                << Eigen::Map<vec3f>(agentBodyNode_->translation().data());
-      break;
-    case KeyEvent::Key::X:
-      controls_(*agentBodyNode_, "moveDown", moveSensitivity, false);
-      LOG(INFO) << "Agent position "
-                << Eigen::Map<vec3f>(agentBodyNode_->translation().data());
-      break;
-    case KeyEvent::Key::Z:
-      controls_(*agentBodyNode_, "moveUp", moveSensitivity, false);
-      LOG(INFO) << "Agent position "
-                << Eigen::Map<vec3f>(agentBodyNode_->translation().data());
-      break;
-    case KeyEvent::Key::O: {
-      if (physicsManager_ != nullptr) {
-        int numObjects = resourceManager_.getNumLibraryObjects();
-        if (numObjects) {
-          int randObjectID = rand() % numObjects;
-          addObject(resourceManager_.getObjectConfig(randObjectID));
-        } else
-          LOG(WARNING) << "No objects loaded, can't add any";
-      } else
-        LOG(WARNING)
-            << "Run the app with --enable-physics in order to add objects";
-    } break;
-    case KeyEvent::Key::P:
-      pokeLastObject();
-      break;
-    case KeyEvent::Key::F:
-      pushLastObject();
-      break;
-    case KeyEvent::Key::K:
-      wiggleLastObject();
-      break;
-    case KeyEvent::Key::U:
-      removeLastObject();
-      break;
-    case KeyEvent::Key::V:
-      invertGravity();
-      break;
-    case KeyEvent::Key::T:
-      // Test key. Put what you want here...
-      torqueLastObject();
-      break;
-    case KeyEvent::Key::I:
-      Magnum::DebugTools::screenshot(GL::defaultFramebuffer,
-                                     "test_image_save.png");
-      break;
-    default:
-      break;
-  }
-  renderCamera_->node().setTransformation(
-      rgbSensorNode_->absoluteTransformation());
-  redraw();
-}
-
-}  // namespace gfx
-}  // namespace esp
